third_party/WebKit/LayoutTests/webaudio/offlineaudiocontext-constructor.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the OfflineAudioContext constructor");

// Make sure we don't crash when giving 0 as number of frames.
shouldThrow("new webkitOfflineAudioContext(1, 0, 44100)");
</script>
</body>
</html>
